refactor: stop inheriting datasets in parsers
