+++ src/otg/datasource/gwas_catalog/study_index.py
@@ -7,16 +7,18 @@ from typing import TYPE_CHECKING
 import pyspark.sql.functions as f
 import pyspark.sql.types as t
 
-from otg.common.spark_helpers import column2camel_case
 from otg.common.utils import parse_efos
 from otg.dataset.study_index import StudyIndex
+from otg.dataset.study_index_gwas_catalog import (
+    StudyIndexGWASCatalog as StudyIndexGWASCatalogDataset,
+)
 
 if TYPE_CHECKING:
     from pyspark.sql import Column, DataFrame
 
 
 @dataclass
-class GWASCatalogStudyIndex(StudyIndex):
+class GWASCatalogStudyIndex:
     """Study index from GWAS Catalog.
 
     The following information is harmonised from the GWAS Catalog:
@@ -202,7 +204,7 @@ class GWASCatalogStudyIndex(StudyIndex):
         )
 
     @staticmethod
-    def _parse_cohorts(raw_cohort: Column) -> Column:
+    def parse_cohorts(raw_cohort: Column) -> Column:
         """Return a list of unique cohort labels from pipe separated list if provided.
 
         Args:
@@ -213,7 +215,7 @@ class GWASCatalogStudyIndex(StudyIndex):
 
         Examples:
         >>> data = [('BioME|CaPS|Estonia|FHS|UKB|GERA|GERA|GERA',),(None,),]
-        >>> spark.createDataFrame(data, ['cohorts']).select(GWASCatalogStudyIndex._parse_cohorts(f.col('cohorts')).alias('parsedCohorts')).show(truncate=False)
+        >>> spark.createDataFrame(data, ['cohorts']).select(GWASCatalogStudyIndex.parse_cohorts(f.col('cohorts')).alias('parsedCohorts')).show(truncate=False)
         +--------------------------------------+
         |parsedCohorts                         |
         +--------------------------------------+
@@ -230,16 +232,16 @@ class GWASCatalogStudyIndex(StudyIndex):
     @classmethod
     def _parse_study_table(
         cls: type[GWASCatalogStudyIndex], catalog_studies: DataFrame
-    ) -> GWASCatalogStudyIndex:
+    ) -> StudyIndexGWASCatalogDataset:
         """Harmonise GWASCatalog study table with `StudyIndex` schema.
 
         Args:
             catalog_studies (DataFrame): GWAS Catalog study table
 
         Returns:
-            GWASCatalogStudyIndex: Parsed and annotated GWAS Catalog study table.
+            StudyIndexGWASCatalogDataset: Parsed and annotated GWAS Catalog study table.
         """
-        return GWASCatalogStudyIndex(
+        return StudyIndexGWASCatalogDataset(
             _df=catalog_studies.select(
                 f.coalesce(
                     f.col("STUDY ACCESSION"), f.monotonically_increasing_id()
@@ -260,7 +262,7 @@ class GWASCatalogStudyIndex(StudyIndex):
                     "backgroundTraitFromSourceMappedIds"
                 ),
             ),
-            _schema=GWASCatalogStudyIndex.get_schema(),
+            _schema=StudyIndexGWASCatalogDataset.get_schema(),
         )
 
     @classmethod
@@ -269,7 +271,7 @@ class GWASCatalogStudyIndex(StudyIndex):
         catalog_studies: DataFrame,
         ancestry_file: DataFrame,
         sumstats_lut: DataFrame,
-    ) -> GWASCatalogStudyIndex:
+    ) -> StudyIndex:
         """Ingests study level metadata from the GWAS Catalog.
 
         Args:
@@ -278,260 +280,12 @@ class GWASCatalogStudyIndex(StudyIndex):
             sumstats_lut (DataFrame): GWAS Catalog summary statistics list.
 
         Returns:
-            GWASCatalogStudyIndex: Parsed and annotated GWAS Catalog study table.
+            StudyIndex: Parsed and annotated GWAS Catalog study table.
         """
         # Read GWAS Catalogue raw data
         return (
             cls._parse_study_table(catalog_studies)
-            ._annotate_ancestries(ancestry_file)
-            ._annotate_sumstats_info(sumstats_lut)
-            ._annotate_discovery_sample_sizes()
+            .annotate_ancestries(ancestry_file)
+            .annotate_sumstats_info(sumstats_lut)
+            .annotate_discovery_sample_sizes()
         )
-
-    def update_study_id(
-        self: GWASCatalogStudyIndex, study_annotation: DataFrame
-    ) -> GWASCatalogStudyIndex:
-        """Update studyId with a dataframe containing study.
-
-        Args:
-            study_annotation (DataFrame): Dataframe containing `updatedStudyId`, `traitFromSource`, `traitFromSourceMappedIds` and key column `studyId`.
-
-        Returns:
-            GWASCatalogStudyIndex: Updated study table.
-        """
-        self.df = (
-            self._df.join(
-                study_annotation.select(
-                    *[
-                        f.col(c).alias(f"updated{c}")
-                        if c not in ["studyId", "updatedStudyId"]
-                        else f.col(c)
-                        for c in study_annotation.columns
-                    ]
-                ),
-                on="studyId",
-                how="left",
-            )
-            .withColumn(
-                "studyId",
-                f.coalesce(f.col("updatedStudyId"), f.col("studyId")),
-            )
-            .withColumn(
-                "traitFromSource",
-                f.coalesce(f.col("updatedtraitFromSource"), f.col("traitFromSource")),
-            )
-            .withColumn(
-                "traitFromSourceMappedIds",
-                f.coalesce(
-                    f.col("updatedtraitFromSourceMappedIds"),
-                    f.col("traitFromSourceMappedIds"),
-                ),
-            )
-            .select(self._df.columns)
-        )
-
-        return self
-
-    def _annotate_ancestries(
-        self: GWASCatalogStudyIndex, ancestry_lut: DataFrame
-    ) -> GWASCatalogStudyIndex:
-        """Extracting sample sizes and ancestry information.
-
-        This function parses the ancestry data. Also get counts for the europeans in the same
-        discovery stage.
-
-        Args:
-            ancestry_lut (DataFrame): Ancestry table as downloaded from the GWAS Catalog
-
-        Returns:
-            GWASCatalogStudyIndex: Slimmed and cleaned version of the ancestry annotation.
-        """
-        ancestry = (
-            ancestry_lut
-            # Convert column headers to camelcase:
-            .transform(
-                lambda df: df.select(
-                    *[f.expr(column2camel_case(x)) for x in df.columns]
-                )
-            ).withColumnRenamed(
-                "studyAccession", "studyId"
-            )  # studyId has not been split yet
-        )
-
-        # Parsing cohort information:
-        cohorts = ancestry_lut.select(
-            f.col("STUDY ACCESSION").alias("studyId"),
-            self._parse_cohorts(f.col("COHORT(S)")).alias("cohorts"),
-        ).distinct()
-
-        # Get a high resolution dataset on experimental stage:
-        ancestry_stages = (
-            ancestry.groupBy("studyId")
-            .pivot("stage")
-            .agg(
-                f.collect_set(
-                    f.struct(
-                        f.col("broadAncestralCategory").alias("ancestry"),
-                        f.col("numberOfIndividuals")
-                        .cast(t.LongType())
-                        .alias("sampleSize"),
-                    )
-                )
-            )
-            .withColumn(
-                "discoverySamples", self._parse_discovery_samples(f.col("initial"))
-            )
-            .withColumnRenamed("replication", "replicationSamples")
-            # Mapping discovery stage ancestries to LD reference:
-            .withColumn(
-                "ldPopulationStructure",
-                self.aggregate_and_map_ancestries(f.col("discoverySamples")),
-            )
-            .drop("initial")
-            .persist()
-        )
-
-        # Generate information on the ancestry composition of the discovery stage, and calculate
-        # the proportion of the Europeans:
-        europeans_deconvoluted = (
-            ancestry
-            # Focus on discovery stage:
-            .filter(f.col("stage") == "initial")
-            # Sorting ancestries if European:
-            .withColumn(
-                "ancestryFlag",
-                # Excluding finnish:
-                f.when(
-                    f.col("initialSampleDescription").contains("Finnish"),
-                    f.lit("other"),
-                )
-                # Excluding Icelandic population:
-                .when(
-                    f.col("initialSampleDescription").contains("Icelandic"),
-                    f.lit("other"),
-                )
-                # Including European ancestry:
-                .when(f.col("broadAncestralCategory") == "European", f.lit("european"))
-                # Exclude all other population:
-                .otherwise("other"),
-            )
-            # Grouping by study accession and initial sample description:
-            .groupBy("studyId")
-            .pivot("ancestryFlag")
-            .agg(
-                # Summarizing sample sizes for all ancestries:
-                f.sum(f.col("numberOfIndividuals"))
-            )
-            # Do arithmetics to make sure we have the right proportion of european in the set:
-            .withColumn(
-                "initialSampleCountEuropean",
-                f.when(f.col("european").isNull(), f.lit(0)).otherwise(
-                    f.col("european")
-                ),
-            )
-            .withColumn(
-                "initialSampleCountOther",
-                f.when(f.col("other").isNull(), f.lit(0)).otherwise(f.col("other")),
-            )
-            .withColumn(
-                "initialSampleCount",
-                f.col("initialSampleCountEuropean") + f.col("other"),
-            )
-            .drop(
-                "european",
-                "other",
-                "initialSampleCount",
-                "initialSampleCountEuropean",
-                "initialSampleCountOther",
-            )
-        )
-
-        parsed_ancestry_lut = ancestry_stages.join(
-            europeans_deconvoluted, on="studyId", how="outer"
-        )
-
-        self.df = self.df.join(parsed_ancestry_lut, on="studyId", how="left").join(
-            cohorts, on="studyId", how="left"
-        )
-        return self
-
-    def _annotate_sumstats_info(
-        self: GWASCatalogStudyIndex, sumstats_lut: DataFrame
-    ) -> GWASCatalogStudyIndex:
-        """Annotate summary stat locations.
-
-        Args:
-            sumstats_lut (DataFrame): listing GWAS Catalog summary stats paths
-
-        Returns:
-            GWASCatalogStudyIndex: including `summarystatsLocation` and `hasSumstats` columns
-        """
-        gwas_sumstats_base_uri = (
-            "ftp://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/"
-        )
-
-        parsed_sumstats_lut = sumstats_lut.withColumn(
-            "summarystatsLocation",
-            f.concat(
-                f.lit(gwas_sumstats_base_uri),
-                f.regexp_replace(f.col("_c0"), r"^\.\/", ""),
-            ),
-        ).select(
-            f.regexp_extract(f.col("summarystatsLocation"), r"\/(GCST\d+)\/", 1).alias(
-                "studyId"
-            ),
-            "summarystatsLocation",
-            f.lit(True).alias("hasSumstats"),
-        )
-
-        self.df = (
-            self.df.drop("hasSumstats")
-            .join(parsed_sumstats_lut, on="studyId", how="left")
-            .withColumn("hasSumstats", f.coalesce(f.col("hasSumstats"), f.lit(False)))
-        )
-        return self
-
-    def _annotate_discovery_sample_sizes(
-        self: GWASCatalogStudyIndex,
-    ) -> GWASCatalogStudyIndex:
-        """Extract the sample size of the discovery stage of the study as annotated in the GWAS Catalog.
-
-        For some studies that measure quantitative traits, nCases and nControls can't be extracted. Therefore, we assume these are 0.
-
-        Returns:
-            GWASCatalogStudyIndex: object with columns `nCases`, `nControls`, and `nSamples` per `studyId` correctly extracted.
-        """
-        sample_size_lut = (
-            self.df.select(
-                "studyId",
-                f.explode_outer(f.split(f.col("initialSampleSize"), r",\s+")).alias(
-                    "samples"
-                ),
-            )
-            # Extracting the sample size from the string:
-            .withColumn(
-                "sampleSize",
-                f.regexp_extract(
-                    f.regexp_replace(f.col("samples"), ",", ""), r"[0-9,]+", 0
-                ).cast(t.IntegerType()),
-            )
-            .select(
-                "studyId",
-                "sampleSize",
-                f.when(f.col("samples").contains("cases"), f.col("sampleSize"))
-                .otherwise(f.lit(0))
-                .alias("nCases"),
-                f.when(f.col("samples").contains("controls"), f.col("sampleSize"))
-                .otherwise(f.lit(0))
-                .alias("nControls"),
-            )
-            # Aggregating sample sizes for all ancestries:
-            .groupBy("studyId")  # studyId has not been split yet
-            .agg(
-                f.sum("nCases").alias("nCases"),
-                f.sum("nControls").alias("nControls"),
-                f.sum("sampleSize").alias("nSamples"),
-            )
-        )
-        self.df = self.df.join(sample_size_lut, on="studyId", how="left")
-        return self
