+++ tests/datasource/gwas_catalog/test_gwas_catalog_study_index.py
@@ -1,22 +1,25 @@
-"""Test StudyIndexGWASCatalog."""
+"""Test GWASCatalogStudyIndex."""
 
 from __future__ import annotations
 
 from pyspark.sql import DataFrame
 
+from otg.dataset.study_index_gwas_catalog import (
+    StudyIndexGWASCatalog as StudyIndexGWASCatalogDataset,
+)
 from otg.datasource.gwas_catalog.study_index import GWASCatalogStudyIndex
 
 
 def test_annotate_discovery_sample_sizes(
-    mock_study_index_gwas_catalog: GWASCatalogStudyIndex,
+    mock_study_index_gwas_catalog: StudyIndexGWASCatalogDataset,
 ) -> None:
     """Test annotate discovery sample sizes."""
     mock_study_index_gwas_catalog.df = mock_study_index_gwas_catalog.df.drop(
         "nCases", "nControls", "nSamples"
     )
     assert isinstance(
-        mock_study_index_gwas_catalog._annotate_discovery_sample_sizes(),
-        GWASCatalogStudyIndex,
+        mock_study_index_gwas_catalog.annotate_discovery_sample_sizes(),
+        StudyIndexGWASCatalogDataset,
     )
 
 
@@ -24,12 +27,12 @@ def test_parse_study_table(sample_gwas_catalog_studies: DataFrame) -> None:
     """Test parse study table."""
     assert isinstance(
         GWASCatalogStudyIndex._parse_study_table(sample_gwas_catalog_studies),
-        GWASCatalogStudyIndex,
+        StudyIndexGWASCatalogDataset,
     )
 
 
 def test_annotate_sumstats(
-    mock_study_index_gwas_catalog: GWASCatalogStudyIndex,
+    mock_study_index_gwas_catalog: StudyIndexGWASCatalogDataset,
     sample_gwas_catalog_harmonised_sumstats_list: DataFrame,
 ) -> None:
     """Test annotate sumstats of GWASCatalogStudyIndex."""
@@ -37,10 +40,10 @@ def test_annotate_sumstats(
         "summarystatsLocation"
     )
     assert isinstance(
-        mock_study_index_gwas_catalog._annotate_sumstats_info(
+        mock_study_index_gwas_catalog.annotate_sumstats_info(
             sample_gwas_catalog_harmonised_sumstats_list
         ),
-        GWASCatalogStudyIndex,
+        StudyIndexGWASCatalogDataset,
     )
 
 
@@ -56,5 +59,5 @@ def test_study_index_from_source(
             sample_gwas_catalog_ancestries_lut,
             sample_gwas_catalog_harmonised_sumstats_list,
         ),
-        GWASCatalogStudyIndex,
+        StudyIndexGWASCatalogDataset,
     )
