+++ tests/datasource/intervals/test_andersson.py
@@ -6,6 +6,7 @@ from pyspark.sql import DataFrame, SparkSession
 
 from otg.common.Liftover import LiftOverSpark
 from otg.dataset.gene_index import GeneIndex
+from otg.dataset.intervals import Intervals
 from otg.datasource.intervals.andersson import IntervalsAndersson
 
 
@@ -30,5 +31,5 @@ def test_andersson_intervals_from_source(
         IntervalsAndersson.parse(
             sample_intervals_andersson, mock_gene_index, liftover_chain_37_to_38
         ),
-        IntervalsAndersson,
+        Intervals,
     )
