+++ tests/datasource/intervals/test_javierre.py
@@ -6,6 +6,7 @@ from pyspark.sql import DataFrame, SparkSession
 
 from otg.common.Liftover import LiftOverSpark
 from otg.dataset.gene_index import GeneIndex
+from otg.dataset.intervals import Intervals
 from otg.datasource.intervals.javierre import IntervalsJavierre
 
 
@@ -30,5 +31,5 @@ def test_javierre_intervals_from_source(
         IntervalsJavierre.parse(
             sample_intervals_javierre, mock_gene_index, liftover_chain_37_to_38
         ),
-        IntervalsJavierre,
+        Intervals,
     )
