+++ tests/datasource/intervals/test_jung.py
@@ -6,6 +6,7 @@ from pyspark.sql import DataFrame, SparkSession
 
 from otg.common.Liftover import LiftOverSpark
 from otg.dataset.gene_index import GeneIndex
+from otg.dataset.intervals import Intervals
 from otg.datasource.intervals.jung import IntervalsJung
 
 
@@ -30,5 +31,5 @@ def test_jung_intervals_from_source(
         IntervalsJung.parse(
             sample_intervals_jung, mock_gene_index, liftover_chain_37_to_38
         ),
-        IntervalsJung,
+        Intervals,
     )
