+++ tests/datasource/intervals/test_thurman.py
@@ -6,6 +6,7 @@ from pyspark.sql import DataFrame, SparkSession
 
 from otg.common.Liftover import LiftOverSpark
 from otg.dataset.gene_index import GeneIndex
+from otg.dataset.intervals import Intervals
 from otg.datasource.intervals.thurman import IntervalsThurman
 
 
@@ -30,5 +31,5 @@ def test_thurman_intervals_from_source(
         IntervalsThurman.parse(
             sample_intervals_thurman, mock_gene_index, liftover_chain_37_to_38
         ),
-        IntervalsThurman,
+        Intervals,
     )
